backend/requirements.txt
google-api-python-client
python-dotenv
requests
beautifulsoup4
lxml
pydantic
nltk
scikit-learn